fix print bug
