+++ frontend/src/components/Download.tsx
@@ -1,10 +1,7 @@
-import { useState } from "react";
-import type { ReactElement, ReactNode, RefObject } from "react";
-import { createPortal } from "react-dom";
+import type { ReactNode, RefObject } from "react";
 import {
   FaBezierCurve,
   FaDownload,
-  FaPrint,
   FaRegImage,
   FaTableCellsLarge,
 } from "react-icons/fa6";
@@ -22,18 +19,13 @@ import {
   downloadSvg,
   downloadTsv,
   downloadTxt,
-  getFilename,
 } from "@/util/download";
-import { sleep } from "@/util/misc";
-import classes from "./Download.module.css";
 
 type Props = {
   /** download filename */
   filename: Filename;
   /** element to render raster image of */
   raster?: RefObject<Element | null>;
-  /** component to print to pdf */
-  print?: ReactElement;
   /** svg element to save */
   vector?: RefObject<SVGSVGElement | null>;
   /** csv/tsv data */
@@ -46,30 +38,16 @@ type Props = {
   children?: ReactNode;
 };
 
-/** app entrypoint element */
-const appElement = document.getElementById("app")!;
-
 /** chart download button */
 const Download = ({
   filename,
   raster,
   vector,
-  print,
   tabular,
   text,
   json,
   children,
 }: Props) => {
-  /** printing state */
-  const [printing, setPrinting] = useState(false);
-
-  /** if printing, render just chart */
-  if (print && printing)
-    return createPortal(
-      <div className={classes.printing}>{print}</div>,
-      document.body,
-    );
-
   return (
     <Popover
       content={
@@ -105,39 +83,6 @@ const Download = ({
                 downloadSvg(vector.current, filename);
               }}
               tooltip="Vector image"
-            />
-          )}
-          {print && (
-            <Button
-              icon={<FaPrint />}
-              text="PDF"
-              onClick={async () => {
-                /** save scroll */
-                const scrollY = window.scrollY;
-                /** save title */
-                const title = document.title;
-                /** set title to suggest pdf filename */
-                document.title = getFilename(filename);
-                /** turn on printing mode */
-                setPrinting(true);
-                /** hide rest of app */
-                appElement.style.display = "none";
-                /** wait for re-render and paint */
-                await sleep();
-                /** open print dialog */
-                window.print();
-                /** turn off printing mode */
-                setPrinting(false);
-                /** restore title */
-                document.title = title;
-                /** re-show rest of app */
-                appElement.style.display = "";
-                /** wait for re-render and paint */
-                await sleep();
-                /** restore scroll */
-                window.scrollTo(0, scrollY);
-              }}
-              tooltip="Print as pdf"
             />
           )}
           {tabular && (
